refactor: :recycle: transfer dataset metadata backend schemas to one file
