+++ server/api/datasets/[datasetId]/metadata/access-rights/index.put.ts
@@ -1,53 +1,4 @@
-import { z } from "zod";
-import FORM_JSON from "@/assets/data/form.json";
-
-const accessTypeOptions = FORM_JSON.datasetAccessTypeOptions.map(
-  (opt) => opt.value,
-);
-
-const accessSchema = z
-  .object({
-    description: z.string().trim().min(1, "Description is required"),
-    type: z
-      .string()
-      .trim()
-      .refine((v) => accessTypeOptions.includes(v), {
-        message: `Access type must be one of: ${accessTypeOptions.join(", ")}`,
-      }),
-    url: z.string().trim().optional(),
-    urlLastChecked: z.string().trim().optional(),
-  })
-  .strict();
-
-const rightsSchema = z
-  .object({
-    identifier: z.string().trim().optional(),
-    identifierScheme: z.string().trim().optional(),
-    identifierSchemeUri: z.string().trim().optional(),
-    licenseText: z.string().trim().optional(),
-    rights: z.string().trim().min(1, "rights is required"),
-    uri: z.string().trim().optional(),
-  })
-  .strict();
-
-const DatasetMetadataAccessRightsSchema = z
-  .object({
-    access: accessSchema,
-    rights: rightsSchema,
-  })
-  .strict()
-  .superRefine((data, ctx) => {
-    if (
-      (data.rights.identifier && !data.rights.identifierScheme) ||
-      (!data.rights.identifier && data.rights.identifierScheme)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message:
-          "Both identifier and identifierScheme must be provided together",
-      });
-    }
-  });
+import { DatasetMetadataAccessRightsSchema } from "@/server/utils/dataset_schemas";
 
 export default defineEventHandler(async (event) => {
   const session = await requireUserSession(event);
