+++ server/api/datasets/[datasetId]/metadata/general-information/index.put.ts
@@ -1,58 +1,4 @@
-import { z } from "zod";
-import FORM_JSON from "@/assets/data/form.json";
-
-const titleTypeOptions = FORM_JSON.datasetTitleTypeOptions.map(
-  (opt) => opt.value,
-);
-const descripTypeOptions = FORM_JSON.datasetDescriptionTypeOptions.map(
-  (opt) => opt.value,
-);
-
-const dateTypeOptions = FORM_JSON.datasetDateTypeOptions.map(
-  (opt) => opt.value,
-);
-
-const aboutSchema = z.object({
-  id: z.string().optional(),
-  date: z.string().min(1, "Date is required"),
-  deleted: z.boolean().optional(),
-  information: z.string().optional(),
-  type: z.string().refine((v) => dateTypeOptions.includes(v), {
-    message: `Date type must be one of: ${dateTypeOptions.join(", ")}`,
-  }),
-});
-
-const descriptionSchema = z.object({
-  id: z.string().optional(),
-  deleted: z.boolean().optional(),
-  description: z.string(),
-  type: z.string().refine((v) => descripTypeOptions.includes(v), {
-    message: `Description type must be one of: ${descripTypeOptions.join(", ")}`,
-  }),
-});
-
-const titleSchema = z.object({
-  id: z.string().optional(),
-  title: z.string().min(1, "Title is required"),
-  deleted: z.boolean().optional(),
-  type: z
-    .string()
-    .optional()
-    .refine(
-      (v) => titleTypeOptions.includes(v as (typeof titleTypeOptions)[number]),
-      {
-        message: `Title type must be one of: ${titleTypeOptions.join(", ")}`,
-      },
-    ),
-});
-
-const DatasetMetadataAboutSchema = z.object({
-  DatasetDate: z.array(aboutSchema).min(1, "At least one date is required"),
-  DatasetDescription: z
-    .array(descriptionSchema)
-    .min(1, "At least Abstract description is required"),
-  DatasetTitle: z.array(titleSchema).min(1, "At least Main title is required"),
-});
+import { DatasetMetadataGeneralInformationSchema } from "~/server/utils/dataset_schemas";
 
 export default defineEventHandler(async (event) => {
   const session = await requireUserSession(event);
@@ -66,7 +12,7 @@ export default defineEventHandler(async (event) => {
 
   // Validate the request body
   const body = await readValidatedBody(event, (b) =>
-    DatasetMetadataAboutSchema.safeParse(b),
+    DatasetMetadataGeneralInformationSchema.safeParse(b),
   );
 
   if (!body.success) {
