+++ server/api/datasets/[datasetId]/metadata/team/index.put.ts
@@ -1,251 +1,4 @@
-import { z } from "zod";
-import FORM_JSON from "@/assets/data/form.json";
-import { isValidORCIDValue, isValidRORValue } from "~/utils/validations";
-
-const nameTypeOptions = FORM_JSON.datasetNameTypeOptions.map(
-  (opt) => opt.value,
-);
-
-const contribTypeOptions = FORM_JSON.datasetContributorTypeOptions.map(
-  (opt) => opt.value,
-);
-
-const funderIdentTypeOptions = FORM_JSON.datasetFunderIdentifierTypeOptions.map(
-  (opt) => opt.value,
-);
-
-// Define the common validation function
-const validateNameIdentifier = (data: any, ctx: z.RefinementCtx) => {
-  if (
-    (data.nameIdentifier && !data.nameIdentifierScheme) ||
-    (!data.nameIdentifier && data.nameIdentifierScheme)
-  ) {
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      message:
-        "Both nameIdentifierScheme and nameIdentifier must be provided together",
-    });
-  }
-
-  if (
-    data.nameIdentifier &&
-    data.nameIdentifierScheme === "ORCID" &&
-    !isValidORCIDValue(data.nameIdentifier)
-  ) {
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      message: "Name identifier must be a valid ORCID value",
-      path: ["nameIdentifier"],
-    });
-  }
-  if (
-    data.nameIdentifier &&
-    data.nameIdentifierScheme === "ROR" &&
-    !isValidRORValue(data.nameIdentifier)
-  ) {
-    ctx.addIssue({
-      code: z.ZodIssueCode.custom,
-      message: "Name identifier must be a valid ROR value",
-      path: ["nameIdentifier"],
-    });
-  }
-};
-
-const affiliationSchema = z
-  .object({
-    affiliation: z.string(),
-    identifier: z.string().optional(),
-    identifierScheme: z.string().optional(),
-    identifierSchemeUri: z.union([z.literal(""), z.string().trim().url()]),
-  })
-  .strict()
-  .superRefine((data, ctx) => {
-    if (
-      (data.identifier && !data.identifierScheme) ||
-      (!data.identifier && data.identifierScheme)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message:
-          "Both identifierScheme and identifier must be provided together",
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierScheme === "ORCID" &&
-      !isValidORCIDValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ORCID value",
-        path: ["identifier"],
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierScheme === "ROR" &&
-      !isValidRORValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ROR value",
-        path: ["identifier"],
-      });
-    }
-  });
-
-// Create base schema for creators and contributors
-const baseCreatorObjectSchema = z
-  .object({
-    id: z.string().optional(),
-    affiliations: z.array(affiliationSchema).optional(),
-    deleted: z.boolean().optional(),
-    familyName: z.string(),
-    givenName: z.string(),
-    local: z.boolean().optional(),
-    nameIdentifier: z.string().optional(),
-    nameIdentifierScheme: z.string().optional(),
-    nameIdentifierSchemeUri: z.union([z.literal(""), z.string().trim().url()]),
-    nameType: z
-      .string()
-      .trim()
-      .min(1, "Name type is required")
-      .refine((v) => nameTypeOptions.includes(v), {
-        message: `Name type must be one of: ${nameTypeOptions.join(", ")}`,
-      }),
-  })
-  .strict();
-
-// Add the nameIdentifier validation to the creator schema
-const creatorSchema = baseCreatorObjectSchema.superRefine(
-  validateNameIdentifier,
-);
-
-// Contributor schema is the same as creator schema but with an additional contributorType field
-const contributorSchema = baseCreatorObjectSchema
-  .extend({
-    contributorType: z
-      .string()
-      .trim()
-      .min(1, "Contributor type is required")
-      .refine((v) => contribTypeOptions.includes(v), {
-        message: `Contributor type must be one of: ${contribTypeOptions.join(", ")}`,
-        path: ["contributorType"],
-      }),
-  })
-  .superRefine(validateNameIdentifier);
-
-const funderSchema = z
-  .object({
-    id: z.string().optional(),
-    name: z.string().min(1, "Name is required"),
-    awardNumber: z.string().trim().optional(),
-    awardTitle: z.string().trim().optional(),
-    awardUri: z.string().trim().optional(),
-    deleted: z.boolean().optional(),
-    identifier: z.string().trim().optional(),
-    identifierSchemeUri: z.string().trim().optional(),
-    identifierType: z
-      .string()
-      .trim()
-      .optional()
-      .refine((v) => funderIdentTypeOptions.includes(v as any), {
-        message: `Identifier type must be one of: ${funderIdentTypeOptions.join(", ")}`,
-        path: ["identifierType"],
-      }),
-    local: z.boolean().optional(),
-  })
-  .strict()
-  .superRefine((data, ctx) => {
-    if (
-      (data.identifierType && !data.identifier) ||
-      (!data.identifierType && data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Both identifierType and identifier must be provided together",
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierType === "ORCID" &&
-      !isValidORCIDValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ORCID value",
-        path: ["identifier"],
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierType === "ROR" &&
-      !isValidRORValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ROR value",
-        path: ["identifier"],
-      });
-    }
-  });
-
-const managingOrgSchema = z
-  .object({
-    name: z.string().trim().min(1, "Name is required"),
-    identifier: z.string().trim().optional(),
-    identifierScheme: z.string().trim().optional(),
-    identifierSchemeUri: z.string().trim().optional(),
-  })
-  .strict()
-  .superRefine((data, ctx) => {
-    if (
-      (data.identifierScheme && !data.identifier) ||
-      (!data.identifierScheme && data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Both identifierType and identifier must be provided together",
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierScheme === "ORCID" &&
-      !isValidORCIDValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ORCID value",
-        path: ["identifier"],
-      });
-    }
-
-    if (
-      data.identifier &&
-      data.identifierScheme === "ROR" &&
-      !isValidRORValue(data.identifier)
-    ) {
-      ctx.addIssue({
-        code: z.ZodIssueCode.custom,
-        message: "Identifier must be a valid ROR value",
-        path: ["identifier"],
-      });
-    }
-  });
-
-const DatasetMetadataTeamSchema = z
-  .object({
-    contributors: z.array(contributorSchema),
-    creators: z.array(creatorSchema),
-    funders: z.array(funderSchema),
-    managingOrganization: managingOrgSchema,
-  })
-  .strict();
+import { DatasetMetadataTeamSchema } from "~/server/utils/dataset_schemas";
 
 export default defineEventHandler(async (event) => {
   const session = await requireUserSession(event);
